+++ test/lib/validation.js
@@ -10,6 +10,14 @@ var
 describe('validation', () => {
 	'use strict';
 
+	describe('#coalesce', () => {
+		it('should properly return the first non-empty value', () => {
+			validation.coalesce(null, null, null, 'value').should.equal('value');
+			validation.coalesce(undefined, [], null, true).should.be.true;
+			should.not.exist(validation.coalesce(undefined, undefined));
+		});
+	});
+
 	describe('#isEmpty', () => {
 		it('should properly detect blank strings', () => {
 			validation.isEmpty('').should.be.true;