+++ history.md
@@ -1,3 +1,9 @@
+# v1.0.1 - 2016/02/18
+
+* Limited possible parameters sent to `http.request` for validation
+* Added `#coalesce` function to `validation` module
+* Added key-api module with tests
+
 # v1.0.0 - 2016/02/16
 
 * Initial release